requirements.txt
alembic==1.7.5
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.1
feedparser==6.0.8
greenlet==1.1.2
Mako==1.1.6
MarkupSafe==2.0.1
mysqlclient==2.1.0
pycparser==2.21
PyMySQL==1.0.2
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
sgmllib3k==1.0.0
six==1.16.0
SQLAlchemy==1.4.27
tornado==6.1
tzdata==2021.5
tzlocal==4.1